Moved application specific implementation to the sample module.
