+++ hawtio-sample-springboot/src/main/java/io/hawt/sample/spring/boot/SampleSpringBootService.java
@@ -4,18 +4,35 @@ import io.hawt.config.ConfigFacade;
 import io.hawt.springboot.HawtPlugin;
 import io.hawt.springboot.HawtioConfiguration;
 import io.hawt.springboot.PluginService;
+import io.hawt.system.ConfigManager;
+import io.hawt.web.AuthenticationFilter;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
 
+import javax.annotation.PostConstruct;
+import javax.servlet.ServletContext;
+
 @SpringBootApplication
 @Import(HawtioConfiguration.class)
 public class SampleSpringBootService {
 
+	@Autowired
+	private ServletContext servletContext;
+
     public static void main(String[] args) {
+		System.setProperty(AuthenticationFilter.HAWTIO_AUTHENTICATION_ENABLED, "false");
 		SpringApplication.run(SampleSpringBootService.class, args);
     }
+
+	@PostConstruct
+	public void init() {
+		final ConfigManager configManager = new ConfigManager();
+		configManager.init();
+		servletContext.setAttribute("ConfigManager", configManager);
+	}
 
 	/**
 	 * Loading an example plugin
